erasure-code: re-implement the example with chunks
The erasure code example plugin is re-implemented using the
encode_chunks() and decode_chunks() methods to show how they
work. Betting on the fact that a plugin implementor is likely to find
this API more straightforward to adapt than the encode() and decode()
helpers which are more convenient from the point of view of the caller
but not from the point of view of the plugin implementor.

Reviewed-By: Christophe Courtaut <christophe.courtaut@gmail.com>
Signed-off-by: Loic Dachary <loic@dachary.org>